react routing to new page
made a very small thing that opens on the login screen, with a button that links to the New Project page. To get rid of the alert thing, remove showAlert() from handleClick function, but i just saw it on gogle adn thought it might come in handy. I'm also not sure how to make the link not blue and underlined, likely a css thing but I can't remember what i did when working on hw5 that well. make sure you pip install react-router-dom if you are making a new react app, it is required to run this