small touchups to fix transition to app.py
if front end does not work for you, try npm install 'react-router-dom@6' in the frontend folder to get the version of router we need to run the site. When doing npm install, also do 'npm install web-vitals' as it is a necessary module for npm run build. the frontend works with display and page transitions, and the backend works with storing and accessing data. Now we must figure out how to communicate between the two. Might be easier said than done, but also might just be easy. Another action item is being able to deploy on heroku. This part seems to be notorious for causing trouble, so I think it would be best to leave at least a day before the deadline to be able to accomplish this.