front to back api connection
these 4 pages have begun working with front end to back end API. we will need project page and hwset page to follow a similar pattern, but adding and replacing fields that are used for navigating the different pages. New user works entirely, but for login we must fiddle with decrypting to allow users to log in to existing accounts. there is a random error on line 26 of PopUp.js, but it does not seem to hinder the app's ability to run.