users complete
creating new users and logging in is now fully complete. it is stored encrypted in the database as well. Now the only question is how we maintain the username throughout separate pages of the project. Additionally, will users remain logged in upon closing the window or will it log them out? Should we have a separate button to log out?